+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/EditRunController.java
@@ -213,9 +213,17 @@ public class EditRunController {
 
     model.put("runDto", mapper.writeValueAsString(Dtos.asDto(run)));
 
+    ObjectNode formConfig = mapper.createObjectNode();
     User user = authorizationManager.getCurrentUser();
-    model.put("userId", user.getId());
-    model.put("userFullName", user.getFullName());
+    formConfig.put("isAdmin", user.isAdmin());
+    if (user.isRunApprover() || user.isAdmin()) {
+      formConfig.put("isRunApprover", true);
+      formConfig.put("userId", user.getId());
+      formConfig.put("userFullName", user.getFullName());
+    } else {
+      formConfig.put("isRunApprover", false);
+    }
+    model.put("formConfig", mapper.writeValueAsString(formConfig));
 
     return new ModelAndView("/WEB-INF/pages/editRun.jsp", model);
   }