+++ sqlstore/src/main/resources/db/migration/V8501__run_approvers.sql
@@ -1,0 +1,4 @@
+ALTER TABLE _Group ADD COLUMN builtIn BOOLEAN NOT NULL DEFAULT FALSE;
+
+INSERT INTO _Group (name, description, builtIn)
+VALUES ('Run Approvers', 'Users allowed to approve run data', TRUE);