+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 Changes:
 
+  * Added "Run Approvers" group. Only users in this group will be allowed to modify "Data Approved"
+    on the Edit Run page
   * Freeze "Parent Alias" column on the bulk Propagate Samples page
   * Allow removing box from freezer on Edit Box page
   * Fixed detailed sample pages loading